Copy over code for fetchngs.
At first, this pipeline was called nf-core/fetchfastq. After initial release we found that this name was already taken by another bioinformatics tool. We took the decision to rename to nf-core/fetchngs.

In this commit I am taking the released code with the new pipeline name and putting it into a single commit.